Move Vmdb::Settings::Walker up a dir
Makes the changes in the next commit easier to grok by making them after
the move, but allows this commit to still keep the build passing.
