+++ lib/vmdb/settings_walker.rb
